backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
redis==5.0.1
anthropic==0.18.1
pydantic==2.5.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pandas==2.2.0
numpy==1.26.3
scipy==1.12.0
requests==2.31.0
yfinance==0.2.36
python-dotenv==1.0.0
